Initial commit
Adding an optional directory for custom components so as to not clutter the shifter root directory.  I left previous component search (via module location) in place. If there is a pre-existing or preferred way of doing this do let me know, I was just patching for my own needs and figured it might help others if not already implemented.